+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- **feat**: `base64` rule as a way to validate whether or not a value is a valid base64 encoded string
+
 ## [9.10.0] - 2024-05-02
 ### Improved
 - **dx**: functions/vAlphaNumSpaces now acts as a typeguard